一旦現在の内容をgitにpush
